refactor(interfaces): move interfaces into separate package
rename imports
